requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
autoflake==2.3.1
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
Brotli==1.1.0
cairocffi==1.7.0
CairoSVG==2.7.1
camelot-py==0.9.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
comm==0.2.2
configparser==3.5.3
cryptography==42.0.8
cssselect2==0.7.0
dataclasses-json==0.6.7
datasets==2.19.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docutils==0.21.2
et-xmlfile==1.1.0
evaluate==0.4.2
excalibur-py==0.4.3
executing==2.0.1
feedparser==6.0.11
filelock==3.14.0
fire==0.6.0
fixit==2.1.0
Flask==1.1.2
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.7
imagesize==1.4.1
img2table==1.2.11
imgkit==1.2.3
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.4
ipython==8.25.0
IronPdf==2024.5.1.2
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.3.7
libcst==1.4.0
llama-index==0.10.46
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.46
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-openai==0.1.10
llama-index-extractors-entity==0.1.2
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.1.5
llama-index-llms-llama-cpp==0.1.3
llama-index-llms-localai==0.1.3
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.22
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-multi-modal-llms-replicate==0.1.3
llama-index-program-lmformatenforcer==0.1.2
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-json==0.1.5
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llama_cpp_python==0.2.78
llamaindex-py-client==0.1.19
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mccabe==0.7.0
minijinja==2.0.1
moreorless==0.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.34.0
opencv-contrib-python==4.10.0.82
opencv-python-headless==4.10.0.82
openparse==0.5.7
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
polars==0.20.31
prompt_toolkit==3.0.46
psutil==5.9.8
psycopg==3.1.19
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydyf==0.10.0
pyflakes==3.2.0
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdf==4.2.0
PyPDF2==2.12.1
pypdfium2==4.30.0
pyphen==0.15.0
pyrate-limiter==3.6.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pdfbox==0.1.8.1
pythonnet==3.0.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
ruff==0.4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sec-edgar-downloader==5.0.2
sentence-transformers==2.7.0
seqeval==1.2.2
setuptools==70.0.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
span-marker==1.5.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.31
stack-data==0.6.3
striprtf==0.0.26
sympy==1.12.1
tabula-py==2.9.3
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
tesseract==0.1.3
text-generation==0.7.0
threadpoolctl==3.5.0
tika==2.6.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
trailrunner==1.4.0
traitlets==5.14.3
transformers==4.40.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==0.16.1
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zopfli==0.2.3
